cleanup_post_metadata function
I miss understood the point of the cleanup of the metadata lines. It was to remove the leading "- " at the beginning of the lines, if it exists. The problem is that it was too agressive.

I left all the various render functions, incase they are needed later, but currently they just return the input string.

the new cleanup_post_metadata function uses php's built in trim, so that it only trims the lines instead of removing data from inside the strings (for example, if the author has a hyphen in their name).